requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
click==8.1.7
frozenlist==1.4.1
idna==3.6
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
Pygments==2.17.2
python-can==4.3.1
python-can-canine==0.2.5
pyusb==1.2.1
rich==13.7.0
textual==0.47.0
textual-dev==1.4.0
typing_extensions==4.9.0
uc-micro-py==1.0.2
wrapt==1.16.0
yarl==1.9.4
